Teste isAdjacent()

+++ trab_algoritmos_em_grafos/src/trab_algoritmos_em_grafos/Aresta.java
@@ -1,15 +1,15 @@
 package trab_algoritmos_em_grafos;
 
 public class Aresta<TIPO> {
-    private Double peso;
+    private int peso;
     private Vertice<TIPO> inicio;
     private Vertice<TIPO> fim;
 
     //Getters e Setters
-    public Double getPeso() {
+    public int getPeso() {
         return peso;
     }
-    public void setPeso(Double peso) {
+    public void setPeso(int peso) {
         this.peso = peso;
     }
     public Vertice<TIPO> getFim() {
@@ -26,7 +26,7 @@ public class Aresta<TIPO> {
     }
     
     //Construtor
-    public Aresta(Double peso, Vertice<TIPO> inicio, Vertice<TIPO> fim){
+    public Aresta(int peso, Vertice<TIPO> inicio, Vertice<TIPO> fim){
         this.peso = peso;
         this.inicio = inicio;
         this.fim = fim;
